+++ cde/lib/DtSvc/DtUtil1/ActionDb.c
@@ -1063,7 +1063,7 @@ _DtActionConverter(DtDtsDbField * fields,
 
 		if ((ThisBit = _DtActFieldBitLookup(XrmQuarkToString(fields[i].fieldName))) != 0)
 		{
-			switch ((int)(fieldVal=_DtActValidateFieldValue(ThisBit,
+			switch ((intptr_t)(fieldVal=_DtActValidateFieldValue(ThisBit,
                                 fields[i].fieldValue, fields[0].fieldValue,
                                 fileName)))
 			{
